Updated to v3.1 (Damage Box Overhaul)
Made the damage box more dynamic, now it changes color when healing (green) or damaging (red).

Also added an alternative position next to the crosshair as an option.

Also included the old version if you prefer it.

(BUG) If you play medic in mvm i reccomend you to use the disable option, because it lags, but works fine for everything else.